+++ src/native/java/lang/util/concurrent/atomic/AtomicLong.cpp
@@ -34,53 +34,53 @@ extern "C" {
 	JNIEXPORT void JNICALL Java_java_util_concurrent_atomic_AtomicLong_lazySet(JNIEnv* env, jobject obj, jlong newValue) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		value->setField("value", sandvik::Object::make(static_cast<uint64_t>(static_cast<int64_t>((static_cast<int64_t>(newValue))))));
+		value->setField("value", sandvik::Object::make(static_cast<uint64_t>(newValue)));
 	}
 
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_getAndSet(JNIEnv* env, jobject obj, jlong newValue) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->getAndSet(static_cast<int64_t>(newValue)));
+		return static_cast<jlong>(value->getAndSet(newValue));
 	}
 	JNIEXPORT jboolean JNICALL Java_java_util_concurrent_atomic_AtomicLong_compareAndSet(JNIEnv* env, jobject obj, jlong expect, jlong update) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jboolean>(value->compareAndSet(static_cast<int64_t>(expect), static_cast<int64_t>(update)));
+		return static_cast<jboolean>(value->compareAndSet(expect, update));
 	}
 
 	JNIEXPORT jboolean JNICALL Java_java_util_concurrent_atomic_AtomicLong_weakCompareAndSet(JNIEnv* env, jobject obj, jlong expect, jlong update) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jboolean>(value->weakCompareAndSet(static_cast<int64_t>(expect), static_cast<int64_t>(update)));
+		return static_cast<jboolean>(value->weakCompareAndSet(expect, update));
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_getAndIncrement(JNIEnv* env, jobject obj) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->getAndAdd((int64_t)1));
+		return value->getAndAdd((int64_t)1);
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_getAndDecrement(JNIEnv* env, jobject obj) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->getAndAdd((int64_t)-1));
+		return value->getAndAdd((int64_t)-1);
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_getAndAdd(JNIEnv* env, jobject obj, jlong delta) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->getAndAdd(static_cast<int64_t>(delta)));
+		return value->getAndAdd(delta);
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_incrementAndGet(JNIEnv* env, jobject obj) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->addAndGet((int64_t)1));
+		return value->addAndGet((int64_t)1);
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_decrementAndGet(JNIEnv* env, jobject obj) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->addAndGet((int64_t)-1));
+		return value->addAndGet((int64_t)-1);
 	}
 	JNIEXPORT jlong JNICALL Java_java_util_concurrent_atomic_AtomicLong_addAndGet(JNIEnv* env, jobject obj, jlong delta) {
 		auto* this_ptr = sandvik::native::getObject(obj);
 		auto value = this_ptr->getField("value");
-		return static_cast<jlong>(value->addAndGet(static_cast<int64_t>(delta)));
+		return value->addAndGet(static_cast<int64_t>(delta));
 	}
 }
